+++ R/licenses.R
@@ -5,8 +5,6 @@
 #' data frame or list, default is `TRUE`.
 #' @param dry_run A logical for toggling a dry run of the request, default is
 #' `FALSE`.
-#' @param rate_limit A logical for toggling automatic rate limiting based on
-#' rate limit headers, default is `FALSE`.
 #' @param api_key A valid OpenAQ API key string, default is `NULL`.
 #'
 #' @return A data frame or a list of the results.
@@ -20,12 +18,10 @@ get_license <- function(
     licenses_id,
     as_data_frame = TRUE,
     dry_run = FALSE,
-    rate_limit = FALSE,
     api_key = NULL) {
   path <- paste("licenses", licenses_id, sep = "/")
   data <- fetch(path,
     dry_run = dry_run,
-    rate_limit = rate_limit,
     api_key = api_key
   )
   if (isTRUE(dry_run)) {
@@ -55,8 +51,6 @@ get_license <- function(
 #' data frame or list, default is `TRUE`.
 #' @param dry_run A logical for toggling a dry run of the request, default is
 #' `FALSE`.
-#' @param rate_limit A logical for toggling automatic rate limiting based on
-#' rate limit headers, default is `FALSE`.
 #' @param api_key A valid OpenAQ API key string, default is `NULL`.
 #'
 #' @return A data frame or a list of the results.
@@ -73,7 +67,6 @@ list_licenses <- function(
     page = NULL,
     as_data_frame = TRUE,
     dry_run = FALSE,
-    rate_limit = FALSE,
     api_key = NULL) {
   param_defs <- list(
     order_by = list(default = NULL, validator = NULL),
@@ -91,7 +84,7 @@ list_licenses <- function(
   path <- "licenses"
   data <- fetch(path, params_list,
     dry_run = dry_run,
-    rate_limit = rate_limit,
+
     api_key = api_key
   )
   if (isTRUE(dry_run)) {