+++ R/locations.R
@@ -5,8 +5,6 @@
 #' data frame or list, default is `TRUE`.
 #' @param dry_run A logical for toggling a dry run of the request, default is
 #' `FALSE`.
-#' @param rate_limit A logical for toggling automatic rate limiting based on
-#' rate limit headers, default is `FALSE`.
 #' @param api_key A valid OpenAQ API key string, default is `NULL`.
 #'
 #' @return A data frame or list of results.
@@ -20,12 +18,11 @@ get_location <- function(
     locations_id,
     as_data_frame = TRUE,
     dry_run = FALSE,
-    rate_limit = FALSE,
     api_key = NULL) {
   path <- paste("locations", locations_id, sep = "/")
   data <- fetch(path,
     dry_run = dry_run,
-    rate_limit = rate_limit,
+
     api_key = api_key
   )
   if (isTRUE(dry_run)) {
@@ -98,8 +95,6 @@ get_location <- function(
 #' data frame or list, default is `TRUE`.
 #' @param dry_run A logical for toggling a dry run of the request, default is
 #' `FALSE`.
-#' @param rate_limit A logical for toggling automatic rate limiting based on
-#' rate limit headers. default is `FALSE`.
 #' @param api_key A valid OpenAQ API key string. default is `NULL`.
 #'
 #' @return A data frame or list of results.
@@ -129,7 +124,6 @@ list_locations <- function(
     page = NULL,
     as_data_frame = TRUE,
     dry_run = FALSE,
-    rate_limit = FALSE,
     api_key = NULL) {
   param_defs <- list(
     bbox = list(
@@ -202,7 +196,7 @@ list_locations <- function(
     path,
     query_params = params_list,
     dry_run = dry_run,
-    rate_limit = rate_limit,
+
     api_key = api_key
   )
   if (isTRUE(dry_run)) {