+++ R/owners.R
@@ -5,8 +5,6 @@
 #' data frame or list, default is `TRUE`.
 #' @param dry_run A logical for toggling a dry run of the request, default is
 #' `FALSE`.
-#' @param rate_limit A logical for toggling automatic rate limiting based on
-#' rate limit headers, default is `FALSE`.
 #' @param api_key A valid OpenAQ API key string, default is `NULL`.
 #'
 #' @return A data frame or a list of the results.
@@ -20,12 +18,11 @@ get_owner <- function(
     owners_id,
     as_data_frame = TRUE,
     dry_run = FALSE,
-    rate_limit = FALSE,
     api_key = NULL) {
   path <- paste("owners", owners_id, sep = "/")
   data <- fetch(path,
     dry_run = dry_run,
-    rate_limit = rate_limit,
+
     api_key = api_key
   )
   if (isTRUE(dry_run)) {
@@ -55,8 +52,6 @@ get_owner <- function(
 #' data frame or list, default is `TRUE`.
 #' @param dry_run A logical for toggling a dry run of the request, default is
 #' `FALSE`.
-#' @param rate_limit A logical for toggling automatic rate limiting based on
-#' rate limit headers, default is `FALSE`.
 #' @param api_key A valid OpenAQ API key string, default is `NULL`.
 #'
 #' @return A data frame or a list of the results.
@@ -73,7 +68,6 @@ list_owners <- function(
     page = NULL,
     as_data_frame = TRUE,
     dry_run = FALSE,
-    rate_limit = FALSE,
     api_key = NULL) {
   param_defs <- list(
     order_by = list(default = NULL, validator = NULL),
@@ -92,7 +86,7 @@ list_owners <- function(
   data <- fetch(path,
     query_params = params_list,
     dry_run = dry_run,
-    rate_limit = rate_limit,
+
     api_key = api_key
   )
   if (isTRUE(dry_run)) {